refactor: LottoManager 클래스 추가 및 잘못된 추상화 제거
